Exemplo de SOAP Finalizado - WS e lado Cliente
