+++ .gitignore
@@ -1,0 +1,3 @@
+/Somador/build/
+/ClienteWS/build/
+/ClienteWS/dist/
