+++ ClienteWS/lib/nblibraries.properties
@@ -1,0 +1,4 @@
+libs.CopyLibs.classpath=\
+    ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
+libs.CopyLibs.displayName=Tarefa CopyLibs
+libs.CopyLibs.prop-version=2.0
