+++ ClienteWS/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\Junior\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
