+++ ClienteWS/src/META-INF/jax-ws-catalog.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<catalog xmlns="urn:oasis:names:tc:entity:xmlns:xml:catalog" prefer="system">
+    <system systemId="http://localhost:8080/Soma/Soma?WSDL" uri="wsdl/localhost_8080/Soma/Soma.wsdl"/>
+    <system systemId="http://localhost:8080/Soma/Soma?xsd=1" uri="wsdl/localhost_8080/Soma/Soma.xsd_1.xsd"/>
+</catalog>
