+++ ClienteWS/xml-resources/web-service-references/Soma/wsdl/localhost_8080/Soma/Soma.wsdl
@@ -1,0 +1,55 @@
+<?xml version='1.0' encoding='UTF-8'?><!-- Published by JAX-WS RI (http://jax-ws.java.net). RI's version is Metro/2.3.2-b608 (trunk-7979; 2015-01-21T12:50:19+0000) JAXWS-RI/2.2.11-b150120.1832 JAXWS-API/2.2.12 JAXB-RI/2.2.12-b141219.1637 JAXB-API/2.2.13-b141020.1521 svn-revision#unknown. --><!-- Generated by JAX-WS RI (http://jax-ws.java.net). RI's version is Metro/2.3.2-b608 (trunk-7979; 2015-01-21T12:50:19+0000) JAXWS-RI/2.2.11-b150120.1832 JAXWS-API/2.2.12 JAXB-RI/2.2.12-b141219.1637 JAXB-API/2.2.13-b141020.1521 svn-revision#unknown. --><definitions xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd" xmlns:wsp="http://www.w3.org/ns/ws-policy" xmlns:wsp1_2="http://schemas.xmlsoap.org/ws/2004/09/policy" xmlns:wsam="http://www.w3.org/2007/05/addressing/metadata" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:tns="http://ws/" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://schemas.xmlsoap.org/wsdl/" targetNamespace="http://ws/" name="Soma">
+<types>
+<xsd:schema>
+<xsd:import namespace="http://ws/" schemaLocation="http://localhost:8080/Soma/Soma?xsd=1"/>
+</xsd:schema>
+</types>
+<message name="hello">
+<part name="parameters" element="tns:hello"/>
+</message>
+<message name="helloResponse">
+<part name="parameters" element="tns:helloResponse"/>
+</message>
+<message name="Somar">
+<part name="parameters" element="tns:Somar"/>
+</message>
+<message name="SomarResponse">
+<part name="parameters" element="tns:SomarResponse"/>
+</message>
+<portType name="Soma">
+<operation name="hello">
+<input wsam:Action="http://ws/Soma/helloRequest" message="tns:hello"/>
+<output wsam:Action="http://ws/Soma/helloResponse" message="tns:helloResponse"/>
+</operation>
+<operation name="Somar">
+<input wsam:Action="http://ws/Soma/SomarRequest" message="tns:Somar"/>
+<output wsam:Action="http://ws/Soma/SomarResponse" message="tns:SomarResponse"/>
+</operation>
+</portType>
+<binding name="SomaPortBinding" type="tns:Soma">
+<soap:binding transport="http://schemas.xmlsoap.org/soap/http" style="document"/>
+<operation name="hello">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="Somar">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+</binding>
+<service name="Soma">
+<port name="SomaPort" binding="tns:SomaPortBinding">
+<soap:address location="http://localhost:8080/Soma/Soma"/>
+</port>
+</service>
+</definitions>
