+++ Somador/lib/nblibraries.properties
