+++ Somador/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=0b8cd798
+build.xml.script.CRC32=5e0c775c
+build.xml.stylesheet.CRC32=5910fda3@1.55.1
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=0b8cd798
+nbproject/build-impl.xml.script.CRC32=1971d39b
+nbproject/build-impl.xml.stylesheet.CRC32=6096d939@1.55.1
