+++ Somador/src/conf/MANIFEST.MF
