--- src/java/control/Consumidor.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package control;
-
-/**
- *
- * @author Junior
- * 
- * CONSUMIDOR DO WEB SERVICE SOAP
- */
-public class Consumidor {
-    
-    private static final String URL_WEBSERVICE = "http://argentumws.caelum.com.br/negociacoes"; //Exemplo do CAELUM
-    
-    public static void main(String[] args){
-        
-        
-        
-    }
-    
-}
